humus-app/humus_code/requirements.txt
pillow
requests
twython
